+++ src/pages/instrumentation.js
@@ -23,7 +23,7 @@ export const query = graphql`
     ) {
       edges {
         node {
-          ...exploreProjectsFields
+          ...projectFields
         }
       }
     }